+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/bulkscanprocessorclient/exceptions/ReadEnvelopeException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscan.orchestrator.exceptions;
+package uk.gov.hmcts.reform.bulkscan.orchestrator.services.bulkscanprocessorclient.exceptions;
 
 public class ReadEnvelopeException extends RuntimeException {
     public ReadEnvelopeException(Throwable cause) {
